move module & mock-data files to sales component
